Changes in act() function
Added a limitation of 1 resp. -1 to the continuous state variable. Added a variable which is counting the number of times the agent changed state.
Added the functionality that an agent weighs the impact of each friend according to his overall number of friends (if someone has only one friend, the impact of this friend is high).